Trimmed down the code base
Remove modules and code that are not needed.
