--- easyunfold/pyvaspwfc/__init__.py
@@ -1,3 +1,0 @@
-"""
-Subpackage of VaspPyWfc
-"""
